fixed problem with implicit cast to float

+++ raycast AR play sound/Assets/Script/EditourCoordinate.cs
@@ -3,6 +3,6 @@
 [Serializable]
 public class EditourCoordinate
 {
-    public float lat;
-    public float lng;
+    public double lat;
+    public double lng;
 }
